Improve testing framework some more

+++ test.sh
@@ -2,6 +2,13 @@
 
 cd "$(dirname "$0")"
 
+# Usage message.
+if [[ $1 == "-h" || $1 == "--help" || $1 == "help" ]]; then
+	echo "usage: test.sh [clean] [FILE.scm ...]"
+	exit 0
+fi
+
+# Clean generated files.
 if [[ $1 == "clean" ]]; then
 	rm -f test.log compare.sh test/out/*.out
 	exit 0
@@ -59,16 +66,22 @@ compare() {
 		echo -n .
 	else
 		echo -n F
-		log "Difference detected on $3"
+		log "Difference detected on $1"
 		diff "$1" "$2" >> $LOG
 
 		if [[ -z $failures ]]; then
 			echo -e "#!/bin/bash\nvimdiff $1 $2" > $CMPSH
 			chmod +x $CMPSH
 		fi
-		failures+=" $3"
+		failures+=" $1"
 		return 1
 	fi
+}
+
+# Print message to say test is being skipped.
+skip_test() {
+	echo -n s
+	log "Skipping test $1 (no ref found)"
 }
 
 # Count the number of arguments.
@@ -76,7 +89,24 @@ count() {
 	echo $#
 }
 
+# Count the total number of tests to be run.
 total=$(count $files)
+
+# Adjust total based on presence of files.
+for f in $files; do
+	base=${f##*/}
+	base=${base%.scm}
+	src=test/src/$base.scm
+	ref=test/ref/$base.ref
+	repl_ref=test/ref/${base}_repl.ref
+
+	if [[ ! -f $src ]]; then
+		echolog "Warning: File $src does not exist"
+		((total--))
+	elif [[ -f $ref && -f $repl_ref ]]; then
+		((total++))
+	fi
+done
 
 # Run the regression tests.
 exitstatus=0
@@ -84,22 +114,40 @@ reg_counter=0
 regression() {
 	failures=
 
-	for src in $files; do
-		base=${src##*/}
+	for f in $files; do
+		base=${f##*/}
+		base=${base%.scm}
+		src=test/src/$base.scm
 		ref=test/ref/$base.ref
 		out=test/out/$base.out
+		repl_ref=test/ref/${base}_repl.ref
+		repl_out=test/out/${base}_repl.out
 
-		if [[ $base == repl_* ]]; then
-			bin/eva < $src > $out 2>&1
-		else
-			bin/eva $src > $out 2>&1
+		if [[ ! -f $src ]]; then
+			continue
 		fi
-		compare $ref $out $src
 
-		((reg_counter++))
+		skip=y
+		if [[ -f $ref ]]; then
+			skip=n
+			bin/eva $src > $out 2>&1
+			compare $ref $out
+			((reg_counter++))
+		fi
+		if [[ -f $repl_ref ]]; then
+			skip=n
+			bin/eva < $src > $repl_out 2>&1
+			compare $repl_ref $repl_out
+			((reg_counter++))
+		fi
+
+		if [[ $skip == "y" ]]; then
+			skip_test $src
+		fi
 	done
 }
 
+# Pluralize the word "test".
 ntests() {
 	if [[ $1 -eq 1 ]]; then
 		echo test
@@ -108,7 +156,14 @@ ntests() {
 	fi
 }
 
+# Print a summary of the test results.
 finish() {
+	if [[ -z $failures ]]; then
+		echo " ok"
+	else
+		echo " FAIL"
+	fi
+
 	all="All "
 	if [[ $reg_counter -ne $total ]]; then
 		all=
@@ -117,10 +172,8 @@ finish() {
 	fi
 
 	if [[ -z $failures ]]; then
-		echo " ok"
 		echolog "$all$reg_counter $(ntests $reg_counter) passed"
 	else
-		echo " FAIL"
 		num=$(count $failures)
 		echolog "$num/$total $(ntests $total) failed"
 		echolog "Failures:"
